feat: adiciona geração inicial da população no algoritmo genético

+++ unidade3/headers/buscas-locais.h
@@ -4,7 +4,7 @@
 #include "../../final/headers/DigrafoMatrizAdj.h"
 
 std::pair<std::vector<int>, int> busca_local(std::pair<std::vector<int>, int> solucao, int tipo_busca, int heuristica, const DigrafoMatrizAdj &grafo);
-double calcula_custo(std::vector<int> ordem_vertices, const DigrafoMatrizAdj &grafo);
+int calcula_custo(std::vector<int> ordem_vertices, const DigrafoMatrizAdj &grafo);
 
 std::pair<std::vector<int>, int> swap(std::pair<std::vector<int>, int> solucao, int vertice, int swap_pos, const DigrafoMatrizAdj &grafo);
 std::pair<std::vector<int>, int> shift(std::pair<std::vector<int>, int> solucao, int vertice, int shift_pos, const DigrafoMatrizAdj &grafo);
